# specs/001-fullstack-todo-auth/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo App API
  description: REST API for the Full-Stack Todo App with Authentication
  version: 1.0.0
  contact:
    name: Hackathon Todo App

servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server

security:
  - BearerAuth: []

paths:
  /api/health:
    get:
      summary: Health check
      description: Check if the API is running
      operationId: healthCheck
      security: []
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"

  /api/todos:
    get:
      summary: List all todos for authenticated user
      description: Returns all todos belonging to the authenticated user, sorted by creation date (newest first)
      operationId: listTodos
      tags:
        - Todos
      responses:
        '200':
          description: List of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Create a new todo
      description: Creates a new todo for the authenticated user
      operationId: createTodo
      tags:
        - Todos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/todos/{todoId}:
    parameters:
      - name: todoId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: The todo's unique identifier

    get:
      summary: Get a specific todo
      description: Returns a single todo by ID (must belong to authenticated user)
      operationId: getTodo
      tags:
        - Todos
      responses:
        '200':
          description: Todo found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update a todo
      description: Updates an existing todo (must belong to authenticated user)
      operationId: updateTodo
      tags:
        - Todos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete a todo
      description: Deletes a todo (must belong to authenticated user)
      operationId: deleteTodo
      tags:
        - Todos
      responses:
        '204':
          description: Todo deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token issued by Better Auth

  schemas:
    Todo:
      type: object
      required:
        - id
        - title
        - completed
        - userId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Task description
          example: "Buy groceries"
        completed:
          type: boolean
          description: Whether the task is completed
          example: false
        userId:
          type: string
          format: uuid
          description: Owner's user ID
          example: "123e4567-e89b-12d3-a456-426614174000"
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2025-12-27T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2025-12-27T10:30:00Z"

    CreateTodoRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Task description
          example: "Buy groceries"

    UpdateTodoRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Updated task description
          example: "Buy groceries and milk"
        completed:
          type: boolean
          description: Updated completion status
          example: true

    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: "Todo not found"

  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Title cannot be empty"

    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Invalid or expired token"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Todo not found"

tags:
  - name: Todos
    description: Todo management endpoints